+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please refer to the SaltStack [Release Notes](https://docs.saltstack.com/en/deve
 - Add `PyGit2` support
 - Expose `/home/salt/data/logs`
 - Run `salt-master` as `salt` user
+- Add support for setting timezone
 
 **2018.3.2**